Improved: moved general funtion to utils, added schedule card in accordion and updated toggle logic (#282)
