+++ src/locales/en.json
@@ -65,6 +65,7 @@
   "Failed to disable rule group.": "Failed to disable rule group.",
   "Failed to update group.": "Failed to update group.",
   "Failed to update threshold facility.": "Failed to update threshold facility.",
+  "Failed to update sequence for some rules.": "Failed to update sequence for some rules.",
   "Failed to fetch rule information.": "Failed to fetch rule information.",
   "Failed to update rules order.": "Failed to update rules order.",
   "Failed to update config facilities.": "Failed to update config facilities.",
@@ -125,6 +126,7 @@
   "No jobs to update": "No jobs to update",
   "No previous occurrence": "No previous occurrence",
   "No rules found, redirecting to product page": "No rules found, redirecting to product page",
+  "No sequence has been changed.": "No sequence has been changed.",
   "No threshold rule found. Invalid job": "No threshold rule found. Invalid job",
   "No threshold rules found": "No threshold rules found",
   "No time zone found": "No time zone found",
@@ -198,6 +200,7 @@
   "Select the column index for the following information in the uploaded CSV": "Select the column index for the following information in the uploaded CSV",
   "Select run time": "Select run time",
   "Select store": "Select store",
+  "Sequence for rules updated successfully.": "Sequence for rules updated successfully.",
   "Service has been scheduled": "Service has been scheduled",
   "Set safety stock": "Set safety stock",
   "Settings": "Settings",
